+++ src/main/java/com/example/flower_shop/controller/BouquetFlowerController.java
@@ -32,6 +32,8 @@ public class BouquetFlowerController {
 
     @DeleteMapping("/{id}") //fshije te dhena
     public void removeFlowerFromBouquet(@PathVariable Long id) {
-        bouquetFlowerRepository.deleteById(id);
+        if(bouquetFlowerRepository.existsById(id)){
+            bouquetFlowerRepository.deleteById(id);
+        }
     }
 }
