+++ src/main/java/com/example/flower_shop/controller/FlowerController.java
@@ -3,6 +3,7 @@ package com.example.flower_shop.controller;
 import com.example.flower_shop.model.Flower;
 import com.example.flower_shop.repository.FlowerRepository;
 
+import com.example.flower_shop.repository.SupplierRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,10 +14,12 @@ import java.util.List;
 @CrossOrigin(origins = "http://localhost:3000") 
 public class FlowerController {
 
+    private final SupplierRepository supplierRepository;
     private final FlowerRepository flowerRepository;
 
-    FlowerController(FlowerRepository flowerRepository){
+    FlowerController(FlowerRepository flowerRepository, SupplierRepository supplierRepository){
         this.flowerRepository = flowerRepository;
+        this.supplierRepository = supplierRepository;
     }
 
     @GetMapping
@@ -38,7 +41,9 @@ public class FlowerController {
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT) // kthe 204 per delete sepse nuk ka body
     public void deleteFlower(@PathVariable Long id) {
-        flowerRepository.deleteById(id);
+        if(flowerRepository.existsById(id)){
+            flowerRepository.deleteById(id);
+        } 
     }
 
     @PutMapping("/{id}")
